(PUP-4702) Replace RGEN based Type system with immutable classes
This commit changes all RGEN based classes in the Puppet Type
system into plain Ruby classes. Each class is immutable which means
that the instance is completely initialized by its initializer and
that only attr_readers exists to read the properties of the instance.

Each class has a `DEFAULT` constant defined which is the singleton
default for the type that the class represents. Some types have
additional constants such as `PCollectionType::DEFAULT_SIZE` or
`PStringType::NON_EMPTY`.

In addition to creating immutable classes, some performance critical
methods were also moved out of the polymorphic dispatch. This includes
methods like `assignable?`, `instance?` and `generalize`. The methods
are still available on the TypeCalculator to ensure backward
compatibility.